api/requirements.txt
fastapi==0.115.6
pydantic-settings==2.6.1
PyJWT==2.8.0
python-decouple==3.8
sqlalchemy==2.0.36
asyncpg==0.30.0
greenlet==3.1.1
alembic==1.14.0
psycopg2-binary==2.9.10
fastapi[standard]